chatbot/requirements.txt
faiss-cpu==1.7.4
fastapi==0.110.0
uvicorn==0.27.1
numpy<2
sentence-transformers==2.2.2
python-docx==1.1.0
